services/notification-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
pydantic==2.4.2
pydantic-settings==2.0.3
grpcio==1.59.2
grpcio-tools==1.59.2
protobuf==4.24.4
redis==5.0.1
aioredis==2.0.1
websockets==12.0
python-jose==3.3.0
python-socketio==5.10.0